+++ src/global/vars.ts
@@ -4,3 +4,4 @@ import getParam from "../utilities/get-query-params";
 // Get debug mode params from url
 export const debugMode: boolean = (getParam('dmdebug') != null && getParam('dmdebug') != 'false');
 
+export const apiUrl: string = "https://api.dailymotion.com/";