+++ src/utilities/html-entities.ts
@@ -1,0 +1,3 @@
+export default function htmlEntities(str: string): string {
+    return String(str).replace(/&/g, '%26').replace(/=/g, '%3d');
+}